Add sso events
