+++ workos/resources/sso.py
@@ -1,6 +1,7 @@
 from typing import List, Literal, Union
 
 from workos.resources.workos_model import WorkOSModel
+from workos.types.sso.connection import Connection
 from workos.typing.literals import LiteralOrUntyped
 from workos.utils.connection_types import ConnectionType
 
@@ -28,28 +29,15 @@ class ProfileAndToken(WorkOSModel):
     profile: Profile
 
 
-ConnectionState = Literal[
-    "active", "deleting", "inactive", "requires_type", "validating"
-]
-
-
 class ConnectionDomain(WorkOSModel):
     object: Literal["connection_domain"]
     id: str
     domain: str
 
 
-class Connection(WorkOSModel):
+class ConnectionWithDomains(Connection):
     """Representation of a Connection Response as returned by WorkOS through the SSO feature."""
 
-    object: Literal["connection"]
-    id: str
-    organization_id: str
-    connection_type: LiteralOrUntyped[ConnectionType]
-    name: str
-    state: LiteralOrUntyped[ConnectionState]
-    created_at: str
-    updated_at: str
     domains: List[ConnectionDomain]
